added display, insert/update average conversations length by date

+++ db/migrations/0010_add_metrics_average_conversations_length_table.js
@@ -1,7 +1,7 @@
 exports.up = knex =>
   knex.schema.createTableIfNotExists('metrics_conversations_length', (table) => {
     table.increments('id').primary();
-    table.integer('conversations_length');
+    table.decimal('conversations_length');
     table.timestamp('timestamp');
   });
